requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckdb>=0.9.0
boto3>=1.34.0
scipy>=1.11.0
lxml>=4.9.0
html5lib>=1.1
openpyxl>=3.1.0
python-multipart>=0.0.6
Pillow>=10.0.0